requirements.txt
numpy
openai
python-dotenv
langchain_core